+++ contracts/contracts/camelot/CamelotDex.sol
@@ -25,10 +25,10 @@ contract CamelotDex is IDex {
   /// @return amountOut The amount of output tokens received from the swap
   function swap(
     bytes calldata _data,
-    uint _amountIn,
-    uint _amountOutMin,
+    uint256 _amountIn,
+    uint256 _amountOutMin,
     address _to,
-    uint _deadline
+    uint256 _deadline
   ) external returns (uint256 amountOut) {
     (address tokenIn, address tokenOut) = abi.decode(_data, (address, address));
