+++ contracts/contracts/rateX/interfaces/IHelperState.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.24;
 interface IHelperState {
     struct Token {
         address _address;
-        uint decimals;
+        uint256 decimals;
     }
 
     struct PoolInfo {
