+++ contracts/contracts/sushiV2/SushiSwapHelper.sol
@@ -14,7 +14,7 @@ contract SushiSwapHelper is IHelperState {
 
   function getPoolsData(PoolInfo[] memory poolsInfo) external view returns (SushiSwapV2Pool[] memory pools) {
     pools = new SushiSwapV2Pool[](poolsInfo.length);
-    for (uint i = 0; i < poolsInfo.length; i++) {
+    for (uint256 i = 0; i < poolsInfo.length; i++) {
       PoolInfo memory poolInfo = poolsInfo[i];
 
       uint[] memory reserves;
