+++ contracts/contracts/uniswapV2/UniswapV2Helper.sol
@@ -14,7 +14,7 @@ contract UniswapV2Helper is IHelperState {
 
   function getPoolsData(PoolInfo[] memory poolsInfo) external view returns (UniswapV2Pool[] memory pools) {
     pools = new UniswapV2Pool[](poolsInfo.length);
-    for (uint i = 0; i < poolsInfo.length; i++) {
+    for (uint256 i = 0; i < poolsInfo.length; i++) {
       PoolInfo memory poolInfo = poolsInfo[i];
 
       uint[] memory reserves;
